Build AAB within the script

+++ scripts/aab_to_apks.sh
@@ -4,7 +4,7 @@
 set -e
 
 if [[ "$#" -lt 1 ]]; then
-  echo "USAGE: [RELEASE_TYPE]"
+  echo "USAGE: RELEASE_TYPE [BUILD_AAB]"
   exit 1
 fi
 
@@ -14,16 +14,22 @@ if ! which bundletool >/dev/null 2>&1; then
 fi
 
 RELEASE_TYPE=$1
-if [[ "$RELEASE_TYPE" == "" ]]; then
-  RELEASE_TYPE=release
+BUILD_AAB=true
+if [[ "$2" == "false" ]] || [[ "$2" == "0" ]]; then
+  BUILD_AAB=false
 fi
-default_name="app-${RELEASE_TYPE}"
-RELEASE_PATH=./app/${RELEASE_TYPE}
-TMP_PATH=tmp
-AAB_PATH=${RELEASE_PATH}/${default_name}.aab
-APKS_PATH=${RELEASE_PATH}/${default_name}.apks
-APK_PATH=${RELEASE_PATH}/${default_name}-universal.apks
-APK_FILE_PATH=${RELEASE_PATH}/${default_name}.apk
+
+APP_VERSION="v$(grep -m1 versionName ./app/build.gradle | awk -F \" '{print $2}')"
+APP_NAME="AppManager_${APP_VERSION}"
+DEFAULT_NAME="app-${RELEASE_TYPE}"
+
+RELEASE_PATH="app/build/outputs/bundle/${RELEASE_TYPE}"
+TMP_PATH="tmp"
+
+AAB_PATH="${RELEASE_PATH}/${DEFAULT_NAME}.aab"
+APKS_PATH="${RELEASE_PATH}/${DEFAULT_NAME}.apks"
+APK_PATH="${RELEASE_PATH}/${DEFAULT_NAME}-universal.apks"
+APK_FILE_PATH="${RELEASE_PATH}/${DEFAULT_NAME}.apk"
 
 SUPPORTED_LANGUAGES=(ar bn de en es fa fr hi in it ja nb pl pt ru tr uk vi zh)
 SUPPORTED_DPIS=(ldpi mdpi tvdpi hdpi xhdpi xxhdpi xxxhdpi)
@@ -37,14 +43,21 @@ SUPPORTED_ARCHS=(armeabi_v7a arm64_v8a x86 x86_64)
 source ./scripts/KeyStore.sh
 
 if [[ "${KEYSTORE}" == "" ]]; then
-  read -rsp "KeyStore file: " KEYSTORE
+  read -rp "KeyStore file: " KEYSTORE
   echo
   read -rsp "KeyStore pass: " KEYSTORE_PASS
   echo
-  read -rsp "Key alias: " KEY_ALIAS
+  read -rp "Key alias: " KEY_ALIAS
   echo
   read -rsp "Key alias pass: " KEY_ALIAS_PASS
   echo
+fi
+
+if [[ "$BUILD_AAB" == true ]]; then
+  if [[ -f "${AAB_PATH}" ]]; then
+    rm "${AAB_PATH}"
+  fi
+  ./gradlew "app:bundle$(tr '[:lower:]' '[:upper:]' <<<${RELEASE_TYPE:0:1})${RELEASE_TYPE:1}"
 fi
 
 if [[ -f ${AAB_PATH} ]]; then
@@ -75,13 +88,17 @@ for dpi in "${SUPPORTED_DPIS[@]}"; do
   mv "base-${dpi}.apk" "config.${dpi}.apk"
 done
 for arch in "${SUPPORTED_ARCHS[@]}"; do
-  mv "base-${arch}.apk" "config.${arch}.apk"
+  if [ -f "base-${arch}.apk" ]; then
+    mv "base-${arch}.apk" "config.${arch}.apk"
+  else
+    echo 2>&1 "base-${arch}.apk: not found."
+  fi
 done
 # Delete rests
 rm ./base-*
 # Make zip
 cd "${lastPWD}"
-zip -j "${RELEASE_PATH}/${default_name}.apks" "${RELEASE_PATH}/${TMP_PATH}/splits"/*
+zip -j "${RELEASE_PATH}/${DEFAULT_NAME}.apks" "${RELEASE_PATH}/${TMP_PATH}/splits"/*
 rm -rf "${RELEASE_PATH:?}/${TMP_PATH}"
 
 # Unzip universal APKS file
@@ -94,3 +111,9 @@ else
 fi
 mv "${RELEASE_PATH}/${TMP_PATH}/universal.apk" "${APK_FILE_PATH}"
 rm -rf "${RELEASE_PATH:?}/${TMP_PATH}"
+
+# Rename files
+mv "${APKS_PATH}" "${RELEASE_PATH}/${APP_NAME}.apks"
+mv "${APK_FILE_PATH}" "${RELEASE_PATH}/${APP_NAME}.apk"
+
+echo "Output generated at $(pwd)/${RELEASE_PATH}"
